REFERENCE: SAA-12 Restrict patch and delete   DESCRIPTION: Created helper for comparing user level to access level.     Added level and self restrictions to user routes.

+++ controllers/userController.js
@@ -1,14 +1,20 @@
-import { getAllUsers, updateUserById, deleteUserById } from '../models/UserModel';
+import { getAllUsers, getUserByUsername, getUserByEmail, updateUserById, deleteUserById } from '../models/UserModel';
+import { compareUserLevels } from '../helpers/compareUserLevels';
 import AppError from '../errors/AppError';
+import bcrypt from 'bcrypt';
 
 const getUsers = async ( req, res, next ) => {
   try {
-    const users = await getAllUsers();
-    res.status(200).send({
-      payload: users
-    });
+    if ( compareUserLevels( req.user.level, 'admin' ) ) {      
+      const users = await getAllUsers();
+      res.status(200).send({
+        payload: users
+      });
+    } else {
+      throw new AppError( 'Only admin or moderator can get all users' );
+    }
   } catch ( error ) {
-    next( new AppError( error.message ) );
+    next( error instanceof AppError ? error : new AppError(error.message) );
   }
 }
 
@@ -22,35 +28,60 @@ const getUserInfo = async ( req, res ) => {
 const updateUser = async ( req, res, next ) => {
   try {
     const id = req.params.userId;
-    const body = { ...req.body };
-    const updatedUser = await updateUserById(id, body);
-    if ( updatedUser ) {
-      res.status(200).send({
-        payload: updatedUser
-      });
+    const isAdmin = compareUserLevels( req.user.level, 'admin' );
+
+    if ( (req.user.id).toString() === id || isAdmin ) {
+      const { username, email, password, level } = req.body;
+      let userUpdate = {};
+
+      if ( password ) {
+        const newPassword = await bcrypt.hash( password, parseInt( process.env.PASSWORD_HASHING_ROUNDS, 10 ) );
+        userUpdate.password = newPassword;
+      }
+      if ( username ) {
+        userUpdate.username = username;
+      }
+      if ( email ) {
+        userUpdate.email = email;
+      }
+      if ( level && isAdmin ) {
+        userUpdate.level = level;
+      }
+      const updatedUser = await updateUserById(id, userUpdate);
+      if ( updatedUser ) {
+        res.status(200).send({
+          payload: updatedUser
+        });
+      } else {
+        throw new AppError( 'User not found!' );
+      }
     } else {
-      throw new AppError( 'User not found!' );
+      throw new AppError( 'Only admin or user himself can update user information!' );
     }
   } catch ( error ) {
-    next( error instanceof AppError ? error : new AppError(error.message) );
+    next( error instanceof AppError ? error : new AppError( error.message ) );
   }
 }
 
 const deleteUser = async ( req, res, next ) => {
   try {
     const id = req.params.userId;
-    const deletedUser = await deleteUserById( id );
-    if ( deletedUser ) {
-      res.status(200).send({
-        payload: {
-          message: 'User succesfully deleted!'
-        }
-      });
+    if ( (req.user.id).toString() === id || compareUserLevels( req.user.level, 'admin' ) ) {
+      const deletedUser = await deleteUserById( id );
+      if ( deletedUser ) {
+        res.status(200).send({
+          payload: {
+            message: 'User succesfully deleted!'
+          }
+        });
+      } else {
+        throw new AppError( 'User not found!' );
+      }
     } else {
-      throw new AppError( 'User not found!' );
+      throw new AppError( 'Only admin or user himself can delete user information!' );
     }
   } catch ( error ) {
-    next( error instanceof AppError ? error : new AppError(error.message) );
+    next( error instanceof AppError ? error : new AppError( error.message) );
   }
 }
